Fixed: input[type=number] empty string always identified as invalid.
Because of Chrome returning an invalid entry as an empty string value (correct as per WhatWG HTML5 Spec, 4.10.7.1.13, value sanitization algorithm), we cannot rely on the value of the field to detect this situation. As such, we'll be checking on the  property, which will indicate if the entry is conforming to the valid input format restriction of the number type input. We cannot do much more than assume that this will not be a desired value and reject it anyway. If the regex specified does indeed validate with the user entered value, we just cannot know.